Revert "refactor to allow module execution"
This reverts commit a0123e951171cb22eddac6c65a23c2d8a28a88dc.
